refactor: move all transaction usecases to transaction's specific folder
