+++ src/data/usecases/transaction/db-create-transaction.ts
@@ -3,8 +3,8 @@ import { TransactionModel } from "src/domain/models/transaction";
 import {
   CreateTransaction,
   CreateTransactionModel,
-} from "../../domain/usecases/transaction/create-transaction";
-import { CreateTransactionRepository } from "../protocols/db/transaction/create-transaction-repository";
+} from "../../../domain/usecases/transaction/create-transaction";
+import { CreateTransactionRepository } from "../../protocols/db/transaction/create-transaction-repository";
 
 export class DbCreateTransaction implements CreateTransaction {
   constructor(private readonly repository: CreateTransactionRepository) {}
