+++ src/data/usecases/transaction/db-find-transaction-by-filters.ts
@@ -2,8 +2,8 @@ import { TransactionModel } from "src/domain/models/transaction";
 import {
   FindTransactionByFilters,
   TransactionFiltersModel,
-} from "../../domain/usecases/transaction/find-transaction-by-filters";
-import { FindTransactionByFiltersRepository } from "../protocols/db/transaction/find-transaction-by-filters-repository";
+} from "../../../domain/usecases/transaction/find-transaction-by-filters";
+import { FindTransactionByFiltersRepository } from "../../protocols/db/transaction/find-transaction-by-filters-repository";
 
 export class DbFindTransactionByFilters implements FindTransactionByFilters {
   constructor(
