+++ src/data/usecases/transaction/db-get-total-amount.ts
@@ -1,6 +1,6 @@
 import { TotalAmountModel } from "src/domain/models/total-amount";
-import { GetTotalAmount } from "../../domain/usecases/transaction/get-total-amount";
-import { GetTotalAmountRepository } from "../protocols/db/transaction/get-total-amount-repository";
+import { GetTotalAmount } from "../../../domain/usecases/transaction/get-total-amount";
+import { GetTotalAmountRepository } from "../../protocols/db/transaction/get-total-amount-repository";
 
 export class DbGetTotalAmount implements GetTotalAmount {
   constructor(private readonly repository: GetTotalAmountRepository) {}
