+++ src/data/usecases/transaction/db-revert-transaction.ts
@@ -1,5 +1,5 @@
-import { RevertTransaction } from "../../domain/usecases/transaction/revert-transaction";
-import { RevertTransactionRepository } from "../protocols/db/transaction/revert-transaction-repository";
+import { RevertTransaction } from "../../../domain/usecases/transaction/revert-transaction";
+import { RevertTransactionRepository } from "../../protocols/db/transaction/revert-transaction-repository";
 
 export class DbRevertTransaction implements RevertTransaction {
   constructor(private readonly repository: RevertTransactionRepository) {}
