+++ src/infra/factories/controller-factory.ts
@@ -2,16 +2,16 @@ import { CreateTransactionController } from "../../presentation/controllers/crea
 import { TransactionRepository } from "../db/sequelize/transaction-repository";
 import { Transaction } from "../db/sequelize/sequelize-instance";
 import { FindTransactionByFiltersController } from "../../presentation/controllers/index-transaction-controller";
-import { DbCreateTransaction } from "../../data/usecases/db-create-transaction";
-import { DbFindTransactionByFilters } from "../../data/usecases/db-find-transaction-by-filters";
-import { DbGetTotalAmount } from "../../data/usecases/db-get-total-amount";
-import { DbRevertTransaction } from "../../data/usecases/db-revert-transaction";
-import { DbDeleteTransaction } from "../../data/usecases/db-delete-transaction";
+import { DbCreateTransaction } from "../../data/usecases/transaction/db-create-transaction";
+import { DbFindTransactionByFilters } from "../../data/usecases/transaction/db-find-transaction-by-filters";
+import { DbGetTotalAmount } from "../../data/usecases/transaction/db-get-total-amount";
+import { DbRevertTransaction } from "../../data/usecases/transaction/db-revert-transaction";
+import { DbDeleteTransaction } from "../../data/usecases/transaction/db-delete-transaction";
 import { GetTotalAmountController } from "../../presentation/controllers/get-total-amount-controller";
 import { RevertTransactionController } from "../../presentation/controllers/revert-transaction-controller";
 import { DeleteTransactionController } from "../../presentation/controllers/delete-transaction-controller";
 import { FindTransactionByIdController } from "../../presentation/controllers/find-transaction-by-id-controller";
-import { DbFindTransactionById } from "../../data/usecases/db-find-transaction-by-id";
+import { DbFindTransactionById } from "../../data/usecases/transaction/db-find-transaction-by-id";
 import { CreateTransactionValidator } from "../../utils/validators";
 
 const transactionRepository = new TransactionRepository(Transaction);
