+++ tests/data/usecases/db-delete-transaction.spec.ts
@@ -1,4 +1,4 @@
-import { DbDeleteTransaction } from "../../../src/data/usecases/db-delete-transaction";
+import { DbDeleteTransaction } from "../../../src/data/usecases/transaction/db-delete-transaction";
 import { DeleteTransactionRepository } from "../../../src/data/protocols/db/transaction/delete-transaction-repository";
 
 const makeSut = (): {
