+++ tests/data/usecases/db-find-transaction-by-id.spec.ts
@@ -1,4 +1,4 @@
-import { DbFindTransactionById } from "../../../src/data/usecases/db-find-transaction-by-id";
+import { DbFindTransactionById } from "../../../src/data/usecases/transaction/db-find-transaction-by-id";
 import { FindTransactionByIdRepository } from "../../../src/data/protocols/db/transaction/find-transaction-by-id-repository";
 import { TransactionModel } from "../../../src/domain/models/transaction";
