+++ tests/data/usecases/db-get-total-amount.spec.ts
@@ -1,4 +1,4 @@
-import { DbGetTotalAmount } from "../../../src/data/usecases/db-get-total-amount";
+import { DbGetTotalAmount } from "../../../src/data/usecases/transaction/db-get-total-amount";
 import { GetTotalAmountRepository } from "../../../src/data/protocols/db/transaction/get-total-amount-repository";
 import { TotalAmountModel } from "../../../src/domain/models/total-amount";
